Rework vlan(4) locking.
Previously the locking of vlan(4) interfaces was not very comprehensive.
Particularly there was very little protection against the destruction of
active vlan(4) interfaces or concurrent modification of a vlan(4)
interface. The former readily produced several different panics.

These changes introduce a new global lock in vlan(4) and broaden the
scope of the existing global rmlock(9) to close these races. vlan(4)
should now be much more resilient to the destruction of active
interfaces and concurrent calls into the configuration path.